fix cast

+++ CMP 205 Unit 1 WPF Application for Comic Book Store scenario/DundeeComicBookStore/Helpers/DBAccessHelper.cs
@@ -297,7 +297,7 @@ namespace DundeeComicBookStore
                     {
                         Name = (string)row["name"],
                         Description = (string)row["description"],
-                        UnitPrice = (uint)(int)row["unitPrice"],
+                        UnitPrice = (decimal)row["unitPrice"],
                         UnitsInStock = (uint)(int)row["stockCount"]
                     };
                     productList.Add(product);
